+++ src/screens/Directory/styles.ts
@@ -1,18 +1,73 @@
-import { StyleSheet } from 'react-native';
+import { Dimensions, StyleSheet } from 'react-native';
 import colors from '@/styles/colors';
 import typography from '@/styles/typography';
 
-const styles = StyleSheet.create({
+const { height } = Dimensions.get('window');
+
+export const styles = StyleSheet.create({
   container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: colors.white,
+    paddingTop: height * 0.07,
+    paddingLeft: 42,
+    paddingRight: 42,
+    paddingBottom: 42,
   },
-  text: {
-    ...typography.mediumRegular,
-    color: colors.gray1,
+
+  headerContainer: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingBottom: 40,
+  },
+
+  headerText: {
+    ...typography.heading5,
+    color: colors.primary,
+    paddingBottom: 30,
+    textAlign: 'center',
+    alignItems: 'center',
+  },
+
+  backButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+
+  backIcon: {
+    width: 24,
+    height: 24,
+  },
+
+  callIcon: {
+    width: 64,
+    height: 64,
+    paddingTop: 45,
+    paddingBottom: 45,
+  },
+
+  content: {
+    gap: 29,
+  },
+
+  detailsContainer: {
+    gap: 10,
+  },
+
+  heading: {
+    ...typography.mediumBold,
+    color: colors.black1,
+  },
+
+  subheading: {
+    ...typography.smallBold,
+    color: colors.gray2,
+  },
+
+  darkText: {
+    ...typography.smallRegular,
+    color: colors.black1,
+  },
+
+  lightText: {
+    ...typography.smallRegular,
+    color: colors.gray3,
   },
 });
-
-export default styles;
